+++ mvc/src/main/java/nextstep/mvc/DispatcherServlet.java
@@ -7,6 +7,7 @@ import jakarta.servlet.http.HttpServletResponse;
 import nextstep.mvc.adapter.HandlerAdapter;
 import nextstep.mvc.adapter.asis.ManualHandlerAdapter;
 import nextstep.mvc.adapter.tobe.AnnotationHandlerAdapter;
+import nextstep.mvc.mapping.HandlerMapping;
 import nextstep.mvc.view.ModelAndView;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
